+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `pictureInPictureActions` property on `<PlayerView>` to configure Picture in Picture actions
   - `PictureInPictureAction.TogglePlayback` to add a play/pause action
   - `PictureInPictureAction.Seek` to add seek forward and seek backward for 10 seconds actions
+- Android: `PictureInPictureConfig.shouldEnterOnBackground` now automatically requests Picture in Picture when the app is backgrounded, matching iOS behavior
 
 ### Changed
 
@@ -18,9 +19,6 @@
 ### Fixed
 
 - Android: `player.setSubtitleTrack()` now correctly disables subtitles when passing `undefined` as a parameter. Calling `player.setSubtitleTrack()` without a parameter is no longer supported
-
-### Fixed
-
 - Android: Improve lifecycle handling to free resources
 
 ## [1.4.0] - 2025-11-06